\name{buffer}
\alias{buffer}
\title{Reading OpenBUGS buffer file}
\description{Reads OpenBUGS buffer file, internally used for intefacing to OpenBUGS.}
\usage{
buffer()
}
\value{Prints the buffer, returns nothing.}
\seealso{\code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}
\keyword{internal}
