\name{currentValues}
\alias{currentValues}
\title{Last sampled values}
\description{This function returns the current (last sampled) values of a variable.}
\usage{
currentValues(nodeLabel)
}
\arguments{
  \item{nodeLabel}{Character vector of length 1, name of a variable in the model.}
}
\value{
Vector of the current (last sampled) values of a variable.
}
\seealso{\code{\link{setValues}}, \code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}

