\name{getChain}
\alias{getChain}
\title{Current chain to be initialized}
\description{This function is intended for internal use only.}
\usage{
getChain()
}
\value{Number of the chain to be initialized next.}
\seealso{\code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}
\keyword{internal}
\concept{chain}
