\name{getNumChains}
\alias{getNumChains}
\title{Number of chains}
\description{This function returns the number of chains being simulated.}
\usage{
getNumChains()
}
\value{
Returns the number of chains from the current simulation.
}
\seealso{\code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}
\concept{chain}
