\name{help.WinBUGS} 
\alias{help.WinBUGS}
\title{WinBUGS documentation} 
\description{Function that open the html version of the OpenBUGS manual} 
\usage{
    help.WinBUGS(browser = getOption("browser"))
}
\arguments{
\item{browser}{the name of the program to be used as hypertext browser. It should be in the PATH, or a full path specified.}
}
\details{
  Not yet available in S-PLUS.
}
\seealso{\code{\link{help.BRugs}}} 
\examples{
\dontrun{
help.WinBUGS() 
}
} 
\keyword{documentation}
\concept{help}
