\name{infoMemory}
\alias{infoMemory}
\title{Show memory usage}
\description{Shows the amount of memory allocated to OpenBUGS}
\usage{
infoMemory()
}
\value{
  Amount of memory allocated to OpenBUGS, in bytes.
}
\seealso{\code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}
