\name{infoModules}
\alias{infoModules}
\title{Loaded modules}
\description{Displays all the modules (dynamic link libraries) in use.}
\usage{
infoModules()
}
\value{Dataframe containing information on all the modules (dynamic link libraries) in use.}
\seealso{\code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}

