\name{loadModule}
\alias{loadModule}
\title{Load a module}
\description{This function loads a module.}
\usage{
loadModule(module)
}
\arguments{
  \item{module}{character, name of the module.}
}
\seealso{See \code{\link{modelModules}} for currently loaded modules. \code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}
