\name{modelDynamic}
\alias{modelEnableDynamic}
\alias{modelDisableDynamic}
\title{Controlling Dynamic Compilation}
\description{These functions are used to enable and disable the feature for dynamic compilation.}
\usage{
modelEnableDynamic()
modelDisableDynamic()
}
\seealso{\code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}

