\name{modelNames}
\alias{modelNames}
\title{Get variable names in model}
\description{This function returns the names of variables contained in the current model.}
\usage{
modelNames()
}
\value{Character vector of names of variables contained in the current model.}
\seealso{\code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}

