\name{modelPrecision}
\alias{modelPrecision}
\title{Setting precision for prec figures}
\description{This function sets the precision to which results are displayed to prec figures.}
\usage{
modelPrecision(prec)
}
\arguments{
  \item{prec}{precision used in the figures}
}
\details{It does not affect the precision of any calculations!}
\seealso{\code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}

