\name{modelSeed}
\alias{modelSetSeed}
\alias{modelGetSeed}
\title{Seed of Random Number Generator}
\description{These functions set/return the seed of the random number generator.}
\usage{
modelSetSeed(newSeed)
modelGetSeed()
}
\arguments{
  \item{newSeed}{a positive, non zero integer}
}
\seealso{\code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}
\concept{update}
\concept{iteration}
\concept{seed}
\concept{random}
