\name{modelSetWD}
\alias{modelSetWD}
\title{Set working directory}
\description{Change the working directory}
\usage{
modelSetWD(dir)
}
\arguments{
  \item{dir}{Directory to change to. A character string} 
}
\details{
  Simply an alias for \code{\link{setwd}} from base R, provided
  to mimic the OpenBUGS script function \code{modelSetWD}. 
}
\keyword{interface}
