\name{rats}
\alias{ratsdata}
\alias{ratsinits}
\docType{data}
\title{ratsdata example}
\description{ratsdata example}
\usage{data(ratsdata)
data(ratsinits)}
\format{
  The list \code{ratsdata} contains data originally taken from section 6 of Gelfand and Smith (1990).
}
\source{
A. Gelfand and A. Smith (1990): Sampling-based Approaches to Calculating Marginal Densities.
\emph{Journal of the American Statistical Association}, 85, 398-409.
}
\keyword{datasets}
