\name{modelFactory}
\alias{modelEnable}
\alias{modelDisable}
\title{Enable and disable factories to create updaters}
\description{These functions enable and disable factories that create updaters.}
\usage{
modelEnable(factory)
modelDisable(factory)
}
\arguments{
  \item{factory}{Character (length 1) name of the factory to be disabled/enabled.}
}
\seealso{\code{\link{BRugs}}, \code{\link{help.WinBUGS}}}
\keyword{interface}
\concept{factory}

